// frontend/src/app/gen-api/schemas/inventory.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Super Course SMS API
 * API documentation for the Super Course SMS services
 * OpenAPI spec version: 1.0.0
 */
import type { InventoryUser } from './inventoryUser';
import type { InventoryCustomer } from './inventoryCustomer';
import type { InventoryItemType } from './inventoryItemType';

/**
 * Inventory model with detailed user and customer information
 */
export interface Inventory {
  id: string;
  /**
   * User details in inventory response
   * @nullable
   */
  user: InventoryUser;
  title: string;
  code: string;
  billing_date: string;
  return_date?: string;
  notes?: string;
  /** Customer details in inventory response */
  customer: InventoryCustomer;
  item_type: InventoryItemType;
  createdAt: string;
  updatedAt: string;
}